# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-Commerce API with MongoDB, JWT Authentication, Product Catalog, User Interactions, and Recommendation System",
        "title": "E-Commerce API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@ecommerce.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/me/account": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft delete current user's account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Delete account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_delivery_dto.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/users/me/interactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get summary of current user's product interactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interactions"
                ],
                "summary": "Get my interactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_domain.UserInteractionSummary"
                        }
                    }
                }
            }
        },
        "/users/me/likes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get products the current user has liked",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interactions"
                ],
                "summary": "Get my liked products",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/me/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Change current user's password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Change password",
                "parameters": [
                    {
                        "description": "Password change",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_delivery_dto.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_delivery_dto.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/users/me/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current user's profile information with detailed profile data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_delivery_dto.ProfileResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update current user's detailed profile information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Profile update",
                        "name": "profile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_delivery_dto.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_delivery_dto.ProfileResponse"
                        }
                    }
                }
            }
        },
        "/users/me/purchases": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get products the current user has purchased",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interactions"
                ],
                "summary": "Get my purchases",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/me/recommendations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get product recommendations based on collaborative filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendations"
                ],
                "summary": "Get personalized product recommendations",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of recommendations",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_domain.RecommendationResponse"
                        }
                    }
                }
            }
        },
        "/users/me/similar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get users with similar interaction patterns",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recommendations"
                ],
                "summary": "Get similar users",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of similar users",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/users/me/views": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get products the current user has viewed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interactions"
                ],
                "summary": "Get my view history",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_PrimeraAizen_e-comm_internal_delivery_dto.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "confirm_password",
                "current_password",
                "new_password"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string",
                    "minLength": 8
                },
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "github_com_PrimeraAizen_e-comm_internal_delivery_dto.ProfileResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_PrimeraAizen_e-comm_internal_delivery_dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_PrimeraAizen_e-comm_internal_delivery_dto.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string",
                    "maxLength": 100
                },
                "country": {
                    "type": "string",
                    "maxLength": 100
                },
                "date_of_birth": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "gender": {
                    "type": "string",
                    "maxLength": 20
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "middle_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "phone": {
                    "type": "string",
                    "maxLength": 50
                },
                "postal_code": {
                    "type": "string",
                    "maxLength": 20
                }
            }
        },
...
        "github_com_PrimeraAizen_e-comm_internal_domain.ProductInteraction": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "interacted_at": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                }
            }
        },
        "github_com_PrimeraAizen_e-comm_internal_domain.ProductRecommendation": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "reason": {
                    "description": "Why recommended",
                    "type": "string"
                },
                "score": {
                    "description": "Similarity/relevance score",
                    "type": "number"
                }
            }
        },
...
        "github_com_PrimeraAizen_e-comm_internal_domain.RecommendationResponse": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "description": "e.g., \"collaborative_filtering\"",
                    "type": "string"
                },
                "generated_at": {
                    "type": "string"
                },
                "recommendations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_domain.ProductRecommendation"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_PrimeraAizen_e-comm_internal_domain.UserInteractionSummary": {
            "type": "object",
            "properties": {
                "liked_products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_domain.ProductInteraction"
                    }
                },
                "purchased_products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_domain.ProductInteraction"
                    }
                },
                "total_likes": {
                    "type": "integer"
                },
                "total_purchases": {
                    "type": "integer"
                },
                "total_views": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "viewed_products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_PrimeraAizen_e-comm_internal_domain.ProductInteraction"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}